chore: split integration tests for speed
